test(dedupe): consolidate order_event_recorder tests (11 → 4 files)
Cluster 07116d624c54: Merge 11 test_order_event_recorder_* files into 4
cohesive modules organized by functional concern:

- test_order_event_recorder.py: init, decision_trace, edge_cases, trade_event
- test_order_event_recorder_rejection.py: rejection, integration_rejection, failure
- test_order_event_recorder_submission.py: broker_rejection, submission_attempt
- test_order_event_recorder_preview.py: preview, success

All 33 tests preserved and passing.
