+++ tests/unit/gpt_trader/features/live_trade/execution/test_order_event_recorder_preview.py
@@ -1,4 +1,4 @@
-"""Tests for OrderEventRecorder.record_preview."""
+"""Tests for OrderEventRecorder preview and success methods."""
 
 from __future__ import annotations
 
@@ -7,11 +7,8 @@ from unittest.mock import MagicMock
 
 import pytest
 
-import gpt_trader.features.live_trade.execution.order_event_recorder as order_event_recorder
-from gpt_trader.core import (
-    OrderSide,
-    OrderType,
-)
+import gpt_trader.features.live_trade.execution.order_event_recorder as recorder_module
+from gpt_trader.core import OrderSide, OrderType
 from gpt_trader.features.live_trade.execution.order_event_recorder import OrderEventRecorder
 
 
@@ -20,8 +17,8 @@ def recorder_mocks(monkeypatch: pytest.MonkeyPatch) -> dict[str, MagicMock]:
     monitoring_logger = MagicMock()
     get_monitoring_logger = MagicMock(return_value=monitoring_logger)
     emit_metric = MagicMock()
-    monkeypatch.setattr(order_event_recorder, "get_monitoring_logger", get_monitoring_logger)
-    monkeypatch.setattr(order_event_recorder, "emit_metric", emit_metric)
+    monkeypatch.setattr(recorder_module, "get_monitoring_logger", get_monitoring_logger)
+    monkeypatch.setattr(recorder_module, "emit_metric", emit_metric)
     return {
         "monitoring_logger": monitoring_logger,
         "get_monitoring_logger": get_monitoring_logger,
@@ -144,3 +141,76 @@ class TestRecordPreview:
         )
         recorder_mocks["emit_metric"].assert_called_once()
         monitoring_logger.log_event.assert_called_once()
+
+
+class TestRecordSuccess:
+    """Tests for record_success method."""
+
+    def test_record_success_logs_order_info(
+        self,
+        order_event_recorder: OrderEventRecorder,
+        order_event_mock_order: MagicMock,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        """Test that successful order is logged."""
+        mock_logger = MagicMock()
+        monkeypatch.setattr(recorder_module, "logger", mock_logger)
+
+        # record_success only logs, doesn't interact with event_store
+        # Just verify it doesn't raise
+        order_event_recorder.record_success(
+            order=order_event_mock_order,
+            symbol="BTC-USD",
+            side=OrderSide.BUY,
+            quantity=Decimal("1.0"),
+            display_price=Decimal("50000"),
+            reduce_only=False,
+        )
+        mock_logger.info.assert_called_once()
+        call_kwargs = mock_logger.info.call_args.kwargs
+        assert call_kwargs["symbol"] == "BTC-USD"
+        assert call_kwargs["reduce_only"] is False
+
+    def test_record_success_handles_reduce_only(
+        self,
+        order_event_recorder: OrderEventRecorder,
+        order_event_mock_order: MagicMock,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        """Test that reduce_only flag is handled."""
+        mock_logger = MagicMock()
+        monkeypatch.setattr(recorder_module, "logger", mock_logger)
+
+        order_event_recorder.record_success(
+            order=order_event_mock_order,
+            symbol="BTC-USD",
+            side=OrderSide.SELL,
+            quantity=Decimal("0.5"),
+            display_price=Decimal("51000"),
+            reduce_only=True,
+        )
+        mock_logger.info.assert_called_once()
+        call_kwargs = mock_logger.info.call_args.kwargs
+        assert call_kwargs["reduce_only"] is True
+
+    def test_record_success_handles_market_price(
+        self,
+        order_event_recorder: OrderEventRecorder,
+        order_event_mock_order: MagicMock,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        """Test that 'market' as display_price is handled."""
+        mock_logger = MagicMock()
+        monkeypatch.setattr(recorder_module, "logger", mock_logger)
+
+        order_event_recorder.record_success(
+            order=order_event_mock_order,
+            symbol="BTC-USD",
+            side=OrderSide.BUY,
+            quantity=Decimal("1.0"),
+            display_price="market",
+            reduce_only=False,
+        )
+        mock_logger.info.assert_called_once()
+        call_kwargs = mock_logger.info.call_args.kwargs
+        assert call_kwargs["price"] == "market"
